removed own controller for systemchangelog and integratedit with systemregister
