+++ src/Authentication/Controllers/SystemRegisterController.cs
@@ -16,6 +16,7 @@ using Altinn.Platform.Authentication.Core.SystemRegister.Models;
 using Altinn.Platform.Authentication.Filters;
 using Altinn.Platform.Authentication.Helpers;
 using Altinn.Platform.Authentication.Model;
+using Altinn.Platform.Authentication.Services;
 using Altinn.Platform.Authentication.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -368,6 +369,39 @@ public class SystemRegisterController : ControllerBase
         }
 
         return Ok(new SystemRegisterUpdateResult(true));
+    }
+
+    /// <summary>
+    /// Gets the change log for a specific system identified by its internal ID.
+    /// </summary>
+    /// <param name="systemId">the system internal id</param>
+    /// <param name="cancellationToken">the cancellation token</param>
+    /// <returns></returns>
+    [HttpGet("vendor/{systemId}/changelog")]
+    [Authorize(Policy = AuthzConstants.POLICY_SCOPE_SYSTEMREGISTER_WRITE)]
+    public async Task<ActionResult<List<SystemChangeLog>>> GetChangeLogAsync(string systemId, CancellationToken cancellationToken = default)
+    {
+        RegisteredSystemResponse registeredSystem = await _systemRegisterService.GetRegisteredSystemInfo(systemId, cancellationToken);
+        if (!AuthenticationHelper.HasWriteAccess(AuthenticationHelper.GetOrgNumber(registeredSystem.Vendor.ID), User))
+        {
+            return Forbid();
+        }
+
+        if (registeredSystem is null)
+        {
+            return NotFound($"System with ID {systemId} not found.");
+        }
+        else
+        {
+            if (registeredSystem.IsDeleted)
+            {
+                return NotFound($"System with ID {systemId} is deleted.");
+            }
+        }
+
+        Guid systemInternalId = registeredSystem.InternalId;
+        var changeLog = await _systemRegisterService.GetChangeLogAsync(systemInternalId, cancellationToken);
+        return Ok(changeLog);
     }
 
     private async Task<ValidationErrorBuilder> ValidateRights(List<Right> rights, CancellationToken cancellationToken)
